Python/requirements.txt
absl-py==0.15.0
aiohttp==3.8.0
aiosignal==1.2.0
alabaster==0.7.12
antlr4-python3-runtime==4.8
anyascii==0.3.0
appdirs==1.4.4
argcomplete==1.12.3
argon2-cffi==21.1.0
asgiref==3.4.1
async-timeout==4.0.0
asynctest==0.13.0
atomicwrites==1.4.0
attrdict==2.0.1
attrs==21.2.0
audioread==2.1.9
Babel==2.8.1
backcall==0.2.0
bce-python-sdk==0.8.64
beautifulsoup4==4.10.0
black==19.10b0
bleach==4.1.0
boto3==1.19.7
botocore==1.22.7
braceexpand==0.1.7
cached-property==1.5.2
cachetools==4.2.4
cfgv==3.3.1
click==8.0.3
colorama==0.4.4
comtypes==1.1.10
conda==4.3.16
conda-env-export==0.5.0
configparser==5.0.2
coqpit==0.0.14
cycler==0.11.0
Cython==0.29.24
cytoolz==0.11.2
debugpy==1.5.1
defusedxml==0.7.1
Distance==0.1.3
distlib==0.3.4
Django==3.2.9
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.17.1
editdistance==0.6.0
entrypoints==0.3
fasttext==0.9.2
filelock==3.3.2
flake8==4.0.1
Flask==2.0.2
Flask-Babel==2.0.0
frozendict==2.0.7
frozenlist==1.2.0
fsspec==2021.10.1
funcy==1.17
future==0.18.2
g2p-en==2.1.0
gdown==4.2.0
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.3.2
google-auth-oauthlib==0.4.6
grpcio==1.41.1
gruut==1.2.3
gruut-ipa==0.9.3
gruut-lang-cs==1.2
gruut-lang-de==1.2
gruut-lang-es==1.2
gruut-lang-fr==1.2.1
gruut-lang-it==1.2
gruut-lang-nl==1.2
gruut-lang-pt==1.2
gruut-lang-ru==1.2
gruut-lang-sv==1.2
h5py==3.5.0
huggingface-hub==0.0.19
hydra-core==1.1.1
identify==2.4.5
imagesize==1.2.0
importlib-metadata==4.8.1
importlib-resources==5.4.0
inflect==5.3.0
iniconfig==1.1.1
ipadic==1.0.0
ipykernel==6.4.2
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==4.3.21
itsdangerous==2.0.1
jedi==0.18.0
jieba==0.42.1
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.1.0
jsonlines==1.2.0
jsonschema==4.1.2
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kaldi-io==0.9.4
kaldi-python-io==1.2.2
kaldiio==2.17.2
kiwisolver==1.3.2
latexcodec==2.0.1
librosa==0.8.0
llvmlite==0.36.0
Markdown==3.3.4
MarkupSafe==2.0.1
marshmallow==3.14.0
matplotlib==3.4.3
matplotlib-inline==0.1.3
mccabe==0.6.1
mecab-python3==1.0.3
megatron-lm==2.2.0
mistune==0.8.4
mpmath==1.2.1
multidict==5.2.0
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nemo-toolkit==1.4.0
nest-asyncio==1.5.1
nltk==3.6.2
nodeenv==1.6.0
notebook==6.4.5
num2words==0.5.10
numba==0.53.0
oauthlib==3.1.1
omegaconf==2.1.1
onnx==1.10.2
OpenCC==1.1.1
opencv-python==4.5.3.56
packaging==21.2
pandas==1.3.4
pandocfilters==1.5.0
pangu==4.0.6.1
parameterized==0.8.1
parso==0.8.2
pathspec==0.9.0
pathtools==0.1.2
pesq==0.0.3
pickleshare==0.7.5
Pillow==8.4.0
pip-api==0.0.23
pipreqs==0.4.11
platformdirs==2.4.1
pluggy==1.0.0
pooch==1.5.2
portalocker==2.3.2
pre-commit==2.17.0
prettytable==3.0.0
prometheus-client==0.12.0
promise==2.3
prompt-toolkit==3.0.21
psutil==5.8.0
py==1.10.0
pyannote.core==4.3
pyannote.database==4.1.1
pyannote.metrics==3.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.8.1
pybtex==0.24.0
pybtex-docutils==1.0.1
pycodestyle==2.8.0
pycosat==0.6.3
pycryptodome==3.13.0
pyDeprecate==0.3.1
pydub==0.25.1
pyflakes==2.4.0
Pygments==2.10.0
pynndescent==0.5.5
pyOpenSSL==21.0.0
pyparsing==2.4.7
pypinyin==0.43.0
pypiwin32==223
pyrsistent==0.18.0
pysbd==0.3.4
PySocks==1.7.1
PySoundFile==0.9.0.post1
pystoi==0.3.3
pytest==6.2.5
pytest-runner==5.3.1
python-crfsuite==0.9.7
python-dateutil==2.8.2
pytorch-lightning==1.4.9
pyttsx3==2.90
pytz==2021.3
pywinpty==1.1.5
pyworld==0.3.0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
rapidfuzz==1.8.2
regex==2021.10.23
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
sacrebleu==2.0.0
sacremoses==0.0.46
scikit-learn==1.0.1
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.8.0
sentencepiece==0.1.96
sentry-sdk==1.4.3
setuptools==58.0.4
shellcheck-py==0.8.0.3
shellingham==1.4.0
shortuuid==1.0.1
simplejson==3.17.5
smmap==5.0.0
snowballstemmer==2.1.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
soupsieve==2.2.1
sox==1.4.1
Sphinx==4.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-bibtex==2.4.1
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
subprocess32==3.5.4
sympy==1.9
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tokenizers==0.10.3
toml==0.10.2
toolz==0.11.2
torch==1.7.1+cu110
torch-stft==0.1.4
torchmetrics==0.6.0
torchvision==0.8.2+cu110
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.12.2
TTS==0.4.1
typed-ast==1.4.3
typer==0.4.0
typing-extensions==3.10.0.2
umap-learn==0.5.1
Unidecode==1.3.2
unidic-lite==1.0.8
virtualenv==20.13.0
visualdl==2.2.3
wandb==0.12.6
wcwidth==0.2.5
webdataset==0.1.62
webencodings==0.5.1
Werkzeug==2.0.2
wget==3.2
wheel==0.37.0
widgetsnbextension==3.5.2
win-inet-pton==1.1.0
wincertstore==0.2
wordninja==2.0.0
wrapt==1.13.3
xlwt==1.3.0
yapf==0.26.0
yarg==0.1.9
yarl==1.7.0
yaspin==2.1.0
youtokentome==1.0.6
zipp==3.6.0